chore: move types to cw_common
