+++ contracts/cosmwasm-vm/cw-common/src/client_response.rs
@@ -1,8 +1,14 @@
+use super::*;
 use crate::errors::CwErrors;
-use crate::types::ClientId;
-use crate::types::ClientType;
+use crate::types::{ClientId, ClientType};
 use cosmwasm_schema::cw_serde;
-use ibc::core::ics02_client::height::Height;
+use ibc::core::ics04_channel::msgs::acknowledgement::Acknowledgement;
+use ibc::core::ics04_channel::packet::Packet;
+use ibc::core::ics04_channel::timeout::TimeoutHeight;
+use ibc::signer::Signer;
+use ibc::timestamp::Timestamp;
+use serde::Deserialize;
+use serde::Serialize;
 use std::str::FromStr;
 #[cw_serde]
 pub struct CreateClientResponse {
@@ -99,13 +105,13 @@ pub struct UpgradeClientResponse {
     client_id: String,
     height: String,
     client_state_commitment: Vec<u8>,
-    consesnus_state_commitment: Vec<u8>,
+    consensus_state_commitment: Vec<u8>,
 }
 
 impl UpgradeClientResponse {
     pub fn new(
         client_state_commitment: Vec<u8>,
-        consesnus_state_commitment: Vec<u8>,
+        consensus_state_commitment: Vec<u8>,
         client_id: String,
         height: String,
     ) -> Self {
@@ -114,7 +120,7 @@ impl UpgradeClientResponse {
                 height,
                 client_id,
                 client_state_commitment,
-                consesnus_state_commitment,
+                consensus_state_commitment,
             }
         }
     }
@@ -122,8 +128,8 @@ impl UpgradeClientResponse {
     pub fn client_state_commitment(&self) -> &[u8] {
         &self.client_state_commitment
     }
-    pub fn consesnus_state_commitment(&self) -> &[u8] {
-        &self.consesnus_state_commitment
+    pub fn consensus_state_commitment(&self) -> &[u8] {
+        &self.consensus_state_commitment
     }
     pub fn get_height(&self) -> &str {
         &self.height
@@ -160,4 +166,89 @@ impl MisbehaviourResponse {
     pub fn client_id(&self) -> Result<ClientId, CwErrors> {
         ClientId::from_str(&self.client_id).map_err(|error| CwErrors::InvalidClientId(error))
     }
+}
+
+#[derive(Debug, Deserialize, Serialize, Clone)]
+pub struct PacketResponse {
+    pub seq_on_a: Sequence,
+    pub port_id_on_a: IbcPortId,
+    pub chan_id_on_a: IbcChannelId,
+    pub port_id_on_b: IbcPortId,
+    pub chan_id_on_b: IbcChannelId,
+    pub data: String,
+    pub timeout_height_on_b: TimeoutHeight,
+    pub timeout_timestamp_on_b: Timestamp,
+}
+
+impl From<PacketResponse> for Packet {
+    fn from(packet: PacketResponse) -> Self {
+        let data = hex::decode(packet.data).unwrap();
+        Packet {
+            seq_on_a: packet.seq_on_a,
+            port_id_on_a: packet.port_id_on_a,
+            chan_id_on_a: packet.chan_id_on_a,
+            port_id_on_b: packet.port_id_on_b,
+            chan_id_on_b: packet.chan_id_on_b,
+            data,
+            timeout_height_on_b: packet.timeout_height_on_b,
+            timeout_timestamp_on_b: packet.timeout_timestamp_on_b,
+        }
+    }
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct PacketDataResponse {
+    pub packet: PacketResponse,
+    pub signer: Signer,
+    pub acknowledgement: Option<Acknowledgement>,
+}
+
+#[cw_serde]
+pub struct OpenConfirmResponse {
+    pub conn_id: String,
+    pub counterparty_client_id: String,
+    pub counterparty_connection_id: String,
+    pub counterparty_prefix: Vec<u8>,
+}
+
+#[cw_serde]
+pub struct OpenTryResponse {
+    pub conn_id: String,
+    pub client_id: String,
+    pub counterparty_client_id: String,
+    pub counterparty_connection_id: String,
+    pub counterparty_prefix: Vec<u8>,
+    pub versions: Vec<u8>,
+    pub delay_period: u64,
+}
+
+impl OpenTryResponse {
+    pub fn new(
+        conn_id: String,
+        client_id: String,
+        counterparty_client_id: String,
+        counterparty_connection_id: String,
+        counterparty_prefix: Vec<u8>,
+        versions: Vec<u8>,
+        delay_period: u64,
+    ) -> Self {
+        Self {
+            conn_id,
+            client_id,
+            counterparty_client_id,
+            counterparty_connection_id,
+            counterparty_prefix,
+            versions,
+            delay_period,
+        }
+    }
+}
+
+#[cw_serde]
+pub struct OpenAckResponse {
+    pub conn_id: String,
+    pub version: Vec<u8>,
+    pub counterparty_client_id: String,
+    pub counterparty_connection_id: String,
+    pub counterparty_prefix: Vec<u8>,
 }
