+++ contracts/cosmwasm-vm/cw-ibc-core/src/ics04_channel/handler.rs
@@ -49,7 +49,7 @@ impl<'a> ValidateChannel for CwIbcCoreContext<'a> {
             Ok(addr) => addr,
             Err(error) => return Err(error),
         };
-        let module_id = types::ModuleId::from(module_id);
+        let module_id = cw_common::types::ModuleId::from(module_id);
         let contract_address = match self.get_route(deps.storage, module_id) {
             Ok(addr) => addr,
             Err(error) => return Err(error),
@@ -411,7 +411,7 @@ impl<'a> ValidateChannel for CwIbcCoreContext<'a> {
             Ok(addr) => addr,
             Err(error) => return Err(error),
         };
-        let module_id = types::ModuleId::from(module_id);
+        let module_id = cw_common::types::ModuleId::from(module_id);
         let contract_address = match self.get_route(deps.storage, module_id) {
             Ok(addr) => addr,
             Err(error) => return Err(error),
