+++ contracts/cosmwasm-vm/cw-ibc-core/src/ics05_port/mod.rs
@@ -4,4 +4,5 @@ pub mod port;
 use cosmwasm_std::Storage;
 use ibc::core::{ics05_port::error::PortError, ics26_routing::context::ModuleId};
 
-use crate::{context::CwIbcCoreContext, ics04_channel::PortId, ContractError};
+use crate::{context::CwIbcCoreContext, ContractError};
+use cw_common::types::PortId;
