+++ contracts/cosmwasm-vm/cw-ibc-core/src/ics26_routing/mod.rs
@@ -1,5 +1,6 @@
 pub mod router;
 use crate::ContractError;
-use crate::{context::CwIbcCoreContext, ics04_channel::StorageKey, types::ModuleId};
+use crate::{context::CwIbcCoreContext, storage_keys::StorageKey};
 use cosmwasm_std::{Addr, Storage};
+use cw_common::types::ModuleId;
 use cw_storage_plus::Map;
